+++ lib/fedex_apis/client.rb
@@ -1,3 +1,6 @@
+# frozen_string_literal: true
+
+require 'active_support/core_ext/object/blank'
 require 'excon'
 require 'json'
 
@@ -10,9 +13,11 @@ require_relative 'resource/rate'
 require_relative 'request/tracking'
 require_relative 'resource/tracking'
 
+require_relative 'request/shipment'
+require_relative 'resource/shipment'
+
 module FedexApis
   class Client
-
     OPTIONS = %i[host client_id client_secret account_number].freeze
 
     attr_accessor :options
@@ -22,31 +27,48 @@ module FedexApis
       load_env_options
     end
 
-    def get_token
+    def token
       response = Request::Token.new(@options).run
       Resource::Token.new(JSON.parse(response.body))
     end
 
     def rate(params)
-      access_token = get_token.access_token
+      access_token = token.access_token
       response = Request::Rate.new(@options, params: params).run(access_token)
       Resource::Rate.new(response.status, response.body)
     end
 
     def track(params)
-      access_token = get_token.access_token
+      access_token = token.access_token
       response = Request::Tracking.new(@options, params: params).run(access_token)
       Resource::Tracking.new(response.status, response.body)
+    end
+
+    def create_shipment(params)
+      access_token = token.access_token
+      response = Request::Shipment.new(@options, params: params).create(access_token)
+      Resource::Shipment.new(response.status, response.body)
+    end
+
+    def validate_shipment(params)
+      access_token = token.access_token
+      response = Request::Shipment.new(@options, params: params).validate(access_token)
+      Resource::Shipment.new(response.status, response.body)
+    end
+
+    def cancel_shipment(tracking_number)
+      access_token = token.access_token
+      response = Request::Shipment.new(@options).cancel(access_token, tracking_number)
+      Resource::Shipment.new(response.status, response.body)
     end
 
     private
 
     def load_env_options
       OPTIONS.each do |option|
-        options[option] ||= ENV["FEDEX_APIS_#{option.upcase}"]
+        options[option] ||= ENV.fetch("FEDEX_APIS_#{option.upcase}", nil)
         raise ArgumentError, "Missing option: #{option}" if options[option].blank?
       end
     end
-
   end
 end